Moved files app capabilities to the capabilities manager
* As an example of the new capabilities API
